Add a moprhing test for versions
